fix: rename module directories
