+++ rs_sandbox/my_crate/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod my_module;
+pub mod modules;
