+++ rs_sandbox/my_crate/src/modules.rs
@@ -1,0 +1,2 @@
+
+pub mod my_module;
